+++ module-batch/src/main/java/inspiration/application/slack/SlackService.java
@@ -1,0 +1,7 @@
+package inspiration.application.slack;
+
+import java.io.File;
+
+public interface SlackService {
+    void sendCsv(File file, String title, String filename);
+}
