+++ module-batch/src/main/java/inspiration/application/tag/TagRankingConfig.java
@@ -1,15 +1,9 @@
 package inspiration.application.tag;
 
-import com.slack.api.Slack;
-import com.slack.api.methods.MethodsClient;
-import com.slack.api.methods.request.files.FilesUploadRequest;
-import com.slack.api.methods.response.files.FilesUploadResponse;
-import inspiration.domain.inspiration_tag.InspirationTag;
+import inspiration.application.slack.SlackService;
 import inspiration.domain.inspiration_tag.InspirationTagRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVPrinter;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
@@ -18,21 +12,10 @@ import org.springframework.batch.core.configuration.annotation.StepBuilderFactor
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.batch.core.step.tasklet.Tasklet;
-import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @Slf4j
 @ConditionalOnProperty(
@@ -46,17 +29,8 @@ public class TagRankingConfig {
     static final String JOB_NAME = "tag-ranking-job";
     private static final String STEP_NAME = "tag-ranking-step";
 
-    private static final DateTimeFormatter FORMATTER_YYYYMMDD = DateTimeFormatter.ofPattern("yyyyMMdd");
-
     private final JobBuilderFactory jobBuilderFactory;
     private final StepBuilderFactory stepBuilderFactory;
-    private final InspirationTagRepository inspirationTagRepository;
-    private final TagGroupService googleSheetTagGroupService;
-
-    @Value("${ygtang.slack.token.bot}")
-    private String botToken;
-    @Value("${ygtang.slack.channel.report}")
-    private String reportChannel;
 
     @Bean
     public Job tagRankingJob() {
@@ -70,103 +44,22 @@ public class TagRankingConfig {
     @JobScope
     public Step tagRankingStep() {
         return stepBuilderFactory.get(STEP_NAME)
-                                 .tasklet(tagRankingTasklet())
+                                 .tasklet(tagRankingTasklet(null, null, null))
                                  .transactionManager(new ResourcelessTransactionManager())
                                  .build();
     }
 
     @Bean
     @StepScope
-    public Tasklet tagRankingTasklet() {
-        return (contribution, chunkContext) -> {
-            List<TagGroup> tagGroups = googleSheetTagGroupService.getTagGroups();
-            List<TagRankingVo> tagRankingVoList = getTagRankingVoList(tagGroups);
-            File csvFile = toCsvFile(tagRankingVoList);
-            sendFileToSlack(csvFile);
-            return RepeatStatus.FINISHED;
-        };
+    public Tasklet tagRankingTasklet(
+            InspirationTagRepository inspirationTagRepository,
+            TagGroupService googleSheetTagGroupService,
+            SlackService slackService
+    ) {
+        return new TagRankingTasklet(
+                inspirationTagRepository,
+                googleSheetTagGroupService,
+                slackService
+        );
     }
-
-    private List<TagRankingVo> getTagRankingVoList(List<TagGroup> tagGroups) {
-        List<InspirationTag> inspirationTags = inspirationTagRepository.findAll();
-
-        Map<String, Set<Long>> contentTagIdSetMap =
-                inspirationTags.stream()
-                               .collect(Collectors.toMap(
-                                       it -> resolveTagName(it.getTag().getContent(), tagGroups),
-                                       it -> Stream.of(it.getTag().getId()).collect(Collectors.toSet()),
-                                       (a, b) -> {
-                                           Set<Long> c = new HashSet<>(a);
-                                           c.addAll(b);
-                                           return c;
-                                       }
-                               ));
-        Map<String, Set<Long>> contentInspirationIdSetMap =
-                inspirationTags.stream()
-                               .collect(Collectors.toMap(
-                                       it -> resolveTagName(it.getTag().getContent(), tagGroups),
-                                       it -> Stream.of(it.getInspiration().getId()).collect(Collectors.toSet()),
-                                       (a, b) -> {
-                                           Set<Long> c = new HashSet<>(a);
-                                           c.addAll(b);
-                                           return c;
-                                       }
-                               ));
-        return contentInspirationIdSetMap.keySet()
-                                         .stream()
-                                         .map(content -> new TagRankingVo(
-                                                 content,
-                                                 contentTagIdSetMap.get(content).size(),
-                                                 contentInspirationIdSetMap.get(content).size()
-                                         ))
-                                         .sorted(Comparator.comparing(TagRankingVo::getInspirationCount)
-                                                           .thenComparing(TagRankingVo::getTagCount)
-                                                           .reversed())
-                                         .filter(it -> it.getTagCount() >= 2 || it.getInspirationCount() >= 2)
-                                         .collect(Collectors.toList());
-    }
-
-    private String resolveTagName(String tagName, List<TagGroup> tagGroups) {
-        for (TagGroup tagGroup : tagGroups) {
-            if (tagGroup.contains(tagName)) {
-                return tagGroup.getName();
-            }
-        }
-        return tagName;
-    }
-
-    private File toCsvFile(List<TagRankingVo> tagRankingVoList) throws IOException {
-        File file = File.createTempFile("tagRanking", "csv");
-        file.deleteOnExit();
-        FileWriter out = new FileWriter(file);
-        CSVFormat csvFormat = CSVFormat.Builder.create()
-                                               .setHeader(TagRankingVo.getCsvHeaders())
-                                               .build();
-        try (CSVPrinter printer = new CSVPrinter(out, csvFormat)) {
-            for (TagRankingVo tagRankingVo : tagRankingVoList) {
-                printer.printRecord(tagRankingVo.getCsvValues());
-            }
-        }
-        return file;
-    }
-
-    private void sendFileToSlack(File file) {
-        LocalDate today = LocalDate.now();
-        try (Slack slack = Slack.getInstance()) {
-            MethodsClient methods = slack.methods(botToken);
-            FilesUploadResponse response = methods.filesUpload(
-                    FilesUploadRequest.builder()
-                                      .channels(Collections.singletonList(reportChannel))
-                                      .title("Tag ranking at " + today)
-                                      .filename("tagRanking_" + today.format(FORMATTER_YYYYMMDD) + ".csv")
-                                      .filetype("text/csv")
-                                      .file(file)
-                                      .build()
-            );
-            log.info("response: {}", response);
-        } catch (Exception e) {
-            log.error("Failed to send file to slack", e);
-        }
-    }
-
 }
